ai-py/Chap7/requirements.txt
gradio==4.32.0
faiss-cpu==1.8.0.post1
dashscope==1.20.12
openai==1.40.8
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-dashscope==0.1.4
llama-index-readers-file==0.1.33
docx2txt==0.8
openpyxl==3.1.5
llama-index-core==0.10.67
llama-index-llms-dashscope==0.1.2
llama-index-readers-dashscope==0.1.2
uvicorn==0.30.6
fastapi==0.114.1
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
simplejson==3.19.3
matplotlib==3.9.2
ragas==0.1.9
langchain_community==0.2.17
alibabacloud_bailian20231229==1.8.2
pandas==2.0.3
alibabacloud_green20220302==2.2.11
oss2==2.19.0